fix: use native blobs in elector renderer
